stfu its working. send mail on user register

+++ server/models/User.js
@@ -1,3 +1,4 @@
+"use strict";
 const db = require("../services/db");
 
 /**
@@ -5,11 +6,11 @@ const db = require("../services/db");
  * @param {*} user
  */
 const register = user => {
-  const { email, username, first, last, password } = user;
+  const { email, username, firstname, lastname, password, confirmationHash } = user;
   return db.query(
-    `INSERT INTO matcha.user (email, username, firstname, lastname, password) 
-      VALUES ($1, $2, $3, $4, $5)`,
-    [email, username, first, last, password]
+    `INSERT INTO matcha.user (email, username, firstname, lastname, password, confirm_email_token) 
+      VALUES ($1, $2, $3, $4, $5, $6) RETURNING id, email, username, firstname, lastname`,
+    [email, username, firstname, lastname, password, confirmationHash]
   );
 };
 
@@ -17,8 +18,7 @@ const register = user => {
  * get a user
  * @param {*} userId
  */
-const getUser = userId =>
-  db.query(`SELECT * FROM matcha.user WHERE id = $1`, [userId]);
+const getUser = userId => db.query(`SELECT * FROM matcha.user WHERE id = $1`, [userId]);
 
 /**
  * get all users paginated
@@ -47,18 +47,19 @@ const updateUser = (user, userId) => {
  * delete a user and its associations with cascade delete
  * @param {*} userId
  */
-const deleteUser = userId =>
-  db.query(`DELETE FROM matcha.user WHERE id = $1`, [userId]);
+const deleteUser = userId => db.query(`DELETE FROM matcha.user WHERE id = $1`, [userId]);
 
-const checkUserExists = (username, email) =>
-  db.query(
-    `SELECT COUNT(*) FROM matcha.user WHERE username = $1 OR email = $2`,
-    [username, email]
-  );
+/**
+ * confirm an user from the confirmation hash.
+ */
+const confirmUser = confirmationHash => {
+    return db.query(`UPDATE matcha.user SET confirm_email_token = null, is_activated = true WHERE confirm_email_token = $1`, [confirmationHash]);
+}
+
 module.exports = {
   register,
   getUser,
   updateUser,
   deleteUser,
-  checkUserExists
+  confirmUser
 };
